Implement 'taskCreator' function
Added in the content from the task creator program for module 2.  Changed the input and force type data accordingly.  
Made modifications to the way the calculator locates and extracts the mean and standard deviation values in 3 functions within the index.js file.
Removed the sub-section of the percentContribution function that calculated and used a modifier value.